refactor: move stdio bridge integration tests to dedicated package
