+++ package.json
@@ -15,7 +15,9 @@
 		"version:sync": "node scripts/sync-version.mjs",
 		"version:patch": "node scripts/bump-version.mjs patch",
 		"version:minor": "node scripts/bump-version.mjs minor",
-		"version:major": "node scripts/bump-version.mjs major"
+		"version:major": "node scripts/bump-version.mjs major",
+		"test": "pnpm -r --if-present test",
+		"test:integration": "pnpm --filter @obsiscripta/integration-tests run test"
 	},
 	"keywords": [],
 	"license": "0-BSD"
