+++ packages/integration-tests/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@obsiscripta/integration-tests",
+  "version": "0.1.4",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "test": "vitest run"
+  },
+  "devDependencies": {
+    "vitest": "^2.1.8"
+  }
+}
